(pub) correcting how products & tickets are deleted (obvious bug on products, improvement on tickets)

+++ apps/pub/modules/store/actions/actions.class.php
@@ -41,8 +41,11 @@ class storeActions extends autoStoreActions
       ->andWhere('bp.product_declination_id = ?', $request->getParameter('id',0))
     ;
     $this->forward404Unless($bps = $q->execute());
-    $bps->delete();
-    $this->getUser()->setFlash('success', __('Your items were successfully removed from your cart.'));
+    if ( $bps->count() > 0 )
+    {
+      $bps->delete();
+      $this->getUser()->setFlash('success', __('Your items were successfully removed from your cart.'));
+    }
     $this->redirect('cart/show');
   }
